Consolidate sequence post components
